+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Fixed movies from Kinox.to not working
 - Add streams to stream result list as they load instead of all at once
 - Speed up video link loading from streamzz.to
+- Don't expand search view on app start
 
 ## [0.5.0] - 2021-09-02
 ### Added